# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List events
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: artistId
          in: query
          required: false
          schema:
            type: string
        - name: venueId
          in: query
          required: false
          schema:
            type: string
        - name: city
          in: query
          required: false
          description: Filter to events whose venue is in this city (case-insensitive)
          schema:
            type: string
      responses:
        "200":
          description: Events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create an event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get an event by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateEvent
      tags: [events]
      summary: Update an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpdate"
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /discovery/tags:
    get:
      operationId: listDiscoveryTags
      tags: [discovery]
      summary: List available discovery tags (genres, moods, themes) with counts
      parameters:
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [genre, mood, theme]
      responses:
        "200":
          description: Discovery tags
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DiscoveryTag"
  /discovery/artists:
    get:
      operationId: discoverArtists
      tags: [discovery]
      summary: Discover artists by mood, genre, and free-text search across analysis
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
        - name: genre
          in: query
          required: false
          schema:
            type: string
        - name: mood
          in: query
          required: false
          schema:
            type: string
        - name: theme
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Matching artists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Artist"
...
  /profiles/{id}/follows:
    get:
      operationId: listFollowedArtists
      tags: [profiles]
      summary: List the artists a user follows
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Followed artists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Artist"
  /profiles/{id}/follows/{artistId}:
    get:
      operationId: getFollowStatus
      tags: [profiles]
      summary: Whether a user follows an artist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: artistId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Follow status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowStatus"
    put:
      operationId: followArtist
      tags: [profiles]
      summary: Follow an artist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: artistId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Follow status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowStatus"
    delete:
      operationId: unfollowArtist
      tags: [profiles]
      summary: Unfollow an artist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: artistId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Follow status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowStatus"
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Artist:
      type: object
      required: [id, artistName, genres, moodTags, themes, verified, createdAt]
      properties:
        id:
          type: string
        profileId:
          type: ["string", "null"]
        artistName:
          type: string
        bio:
          type: ["string", "null"]
        summary:
          type: ["string", "null"]
        genres:
          type: array
          items:
            type: string
        moodTags:
          type: array
          items:
            type: string
        themes:
          type: array
          items:
            type: string
        city:
          type: ["string", "null"]
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        imageUrl:
          type: ["string", "null"]
        spotifyUrl:
          type: ["string", "null"]
        instagramUrl:
          type: ["string", "null"]
        youtubeUrl:
          type: ["string", "null"]
        websiteUrl:
          type: ["string", "null"]
        verified:
          type: boolean
        createdAt:
          type: string
...
    Event:
      type: object
      required: [id, name, eventDate, status, createdAt]
      properties:
        id:
          type: string
        venueId:
          type: ["string", "null"]
        artistId:
          type: ["string", "null"]
        name:
          type: string
        description:
          type: ["string", "null"]
        eventDate:
          type: string
        status:
          type: string
          enum: [upcoming, live, past, cancelled]
        imageUrl:
          type: ["string", "null"]
        ticketUrl:
          type: ["string", "null"]
        createdAt:
          type: string
    EventInput:
      type: object
      required: [name, eventDate]
      properties:
        venueId:
          type: string
        artistId:
          type: string
        name:
          type: string
          minLength: 1
        description:
          type: string
        eventDate:
          type: string
        status:
          type: string
          enum: [upcoming, live, past, cancelled]
        imageUrl:
          type: string
        ticketUrl:
          type: string
    EventUpdate:
      type: object
      properties:
        venueId:
          type: string
        artistId:
          type: string
        name:
          type: string
          minLength: 1
        description:
          type: string
        eventDate:
          type: string
        status:
          type: string
          enum: [upcoming, live, past, cancelled]
        imageUrl:
          type: string
        ticketUrl:
          type: string
...
    DiscoveryTag:
      type: object
      required: [tag, type, count]
      properties:
        tag:
          type: string
        type:
          type: string
        count:
          type: integer
...
    FollowStatus:
      type: object
      required: [following]
      properties:
        following:
          type: boolean
...